% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregateName}
\alias{aggregateName}
\title{Aggregate Interaction Data}
\usage{
aggregateName(
  data,
  aggregate = c("row", "column"),
  counted = FALSE,
  cutoff = 0,
  split.by = "_|[|]"
)
}
\arguments{
\item{data}{Interaction data with rows of cell-cell paris and columns of gene-gene pairs.}

\item{aggregate}{Type of aggregation used:
\itemize{
  \item row, using rownames (Default).
  \item column, using colnames.
  }}

\item{counted}{Binarize data to aggregate. Default is FALSE.}

\item{cutoff}{Cutoff for interactions. Default is 0. Only values that are above this threshold will be considered.}

\item{split.by}{interaction pairs. Default is "_" or "|".}
}
\value{
Returns the aggregated matrix.
}
\description{
Aggregate interaction data by row or column names.
}
\keyword{internal}
