% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asSparse}
\alias{asSparse}
\title{Convert Matrix to Sparse Matrix}
\usage{
asSparse(mat, zero_percent = 0.7)
}
\arguments{
\item{mat}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}
}
\value{
Returns a regular or sparse matrix.
}
\description{
Convert Matrix to Sparse Matrix
}
\keyword{internal}
