% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{calcDensity}
\alias{calcDensity}
\title{Calculate the Density of Feature}
\usage{
calcDensity(x, breaks, drop0 = FALSE)
}
\arguments{
\item{breaks}{A vector giving the breakpoints between histogram cells. See details \code{\link[graphics]{hist}}.}

\item{drop0}{Whether to drop all the zeros. Default is FALSE.}
}
\value{
Returns the density.
}
\description{
Calculate the density of an input feature.
}
\keyword{internal}
