% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{calcEMD}
\alias{calcEMD}
\title{Calculate the Earth Mover's Distance Similarity for Interaction Data}
\usage{
calcEMD(
  data,
  idents,
  ligand,
  receptor,
  nbins = 100,
  weighted = TRUE,
  min_pct = 0,
  max_pct = 1,
  drop_zero = FALSE
)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{idents}{Cell type identity in characters or factors.}

\item{ligand}{Name or index for ligand.}

\item{receptor}{Name or index for receptor.}

\item{nbins}{Number of bins to use for Earth Mover's Distance calculation. Default is 100.}

\item{weighted}{Logic, compute weighted Earth Mover's Distance which will give higher weight/penalty on lowly expressed genes. Default is TRUE.}

\item{min_pct}{Values below this percentile will be replaced with this percentile's value. Default is 0.}

\item{max_pct}{Values above this percentile will be replaced with this percentile's value. Default is 1.}

\item{drop_zero}{Whether to drop all the zeros in the data. Default is FALSE.}
}
\value{
Returns Earth Mover's Distance matrix.
}
\description{
Calculate the Earth Mover's Distance Similarity for interaction data. The similarity is defined as 1-normalized Earth Mover's Distance.
}
\keyword{internal}
