% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calcStrength}
\alias{calcStrength}
\title{Calculate the Strength for Multiple Interacting Pair.}
\usage{
calcStrength(ligand_mat, receptor_mat, f = c("+", "*"), threshold = 0)
}
\arguments{
\item{ligand_mat}{Ligand expression matrix.}

\item{receptor_mat}{Receptor expression matrix.}

\item{f}{Functions used to calculate pair strength.
\itemize{
  \item +, using sum of the pairs (Default).
  \item *, using product of the pairs.
  }}

\item{threshold}{Threshold used for interaction products. Default is 0, therefore interaction product (ligand*receptor) equal or below 0 will be removed.}
}
\value{
Returns a matrix with calculated interacting strength.
}
\description{
Calculate the strength for multiple interacting pair. See details \code{\link[scMatchmaker]{calcStrength.single}}.
}
\keyword{internal}
