% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{calcStrength.single}
\alias{calcStrength.single}
\title{Calculate the Strength for Each Interacting Pair.}
\usage{
calcStrength.single(ligand, receptor, f = c("+", "*"), threshold = 0)
}
\arguments{
\item{ligand}{A ligand expression data.}

\item{receptor}{A receptor expression data.}

\item{f}{Functions used to calculate pair strength.
\itemize{
  \item +, using sum of the pairs (Default).
  \item *, using product of the pairs.
  }}

\item{threshold}{Threshold used for interaction products. Default is 0, therefore interaction product (ligand*receptor) equal or below 0 will be removed.}
}
\value{
Returns calculated interacting strength.
}
\description{
Calculate the strength for each interacting pair.
}
\keyword{internal}
