% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellphonedb}
\alias{cellphonedb}
\title{CellPhoneDB Database}
\format{
A data frame with 1574 rows (ligand-receptor pairs) and 15 columns (attributes):
\describe{
  \item{genename_a, genename_b}{interaction partner gene names}
  \item{id_cp_interaction}{CellPhoneDB interaction id}
  \item{partner_a, partner_b}{interaction partner uniprot ids}
  \item{protein_name_a, protein_name_b}{interaction partner protein names}
  \item{annotation_strategy}{annotation strategy}
  \item{source}{annotation strategy}
  \item{subunit_a_1, subunit_a_2, subunit_a_3}{interaction partner a complex subunits}
  \item{subunit_b_1, subunit_b_2, subunit_b_3}{interaction partner b complex subunits} 
}
}
\source{
\url{https://www.cellphonedb.org/}
}
\usage{
cellphonedb
}
\description{
A preloaded \href{https://www.cellphonedb.org/}{CellPhoneDB} database of ligand-receptor interaction pairs.
}
\keyword{datasets}
