% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIdents}
\alias{checkIdents}
\title{Check Cell Identity in Data}
\usage{
checkIdents(data, idents)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{idents}{Cell type identity in characters or factors.}
}
\value{
Returns the checked identity.
}
\description{
Check if cell identities provided matched to the single cell data and reorder them. Any identity name with "_: will be subsituted as ".".
}
\keyword{internal}
