% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{convertData}
\alias{convertData}
\title{Convert Interaction Matrix}
\usage{
convertData(interact.strength, interact.pval)
}
\arguments{
\item{interact.strength}{Interaction strength matrix.}

\item{interact.pval}{Interaction p-value matrix.}
}
\value{
Returns a molten data frame.
}
\description{
Convert interaction matrix into a molten data frame. Interactions with zero strenghth will be removed.
}
\keyword{internal}
