% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Converting}
\alias{Converting}
\title{Convert Interaction Data}
\usage{
Converting(object, selected = TRUE)
}
\arguments{
\item{object}{Matchmaker object.}

\item{selected}{Use selected data if calculated. Default is TRUE.}
}
\value{
Returns a ranked list of interactions.
}
\description{
Convert interaction data frame into a molten data frame using \code{\link[reshape2]{melt}}. If there are many interactions the molten data can be very long. 
If \code{\link[scMatchmaker]{Selecting}} was run, it will convert the data in @selected slot Interactions with zero strenghth will be removed.
}
\examples{
\dontrun{
convert.data <- Converting(object)
}
}
