% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decidua.annotation}
\alias{decidua.annotation}
\title{Decidua Single Cell RNA-seq Annotation Metadata}
\format{
A data frame with 3000 rows (cells) and 4 columns (attributes):
\describe{
  \item{Fetus}{fetus ID}
  \item{location}{organ name}
  \item{final_cluster}{cluster ID}
  \item{annotation}{cell type annotation}
}
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6701/}
}
\usage{
decidua.annotation
}
\description{
A subset of decidua scRNA-seq metadata generated by \href{https://www.nature.com/articles/s41586-018-0698-6}{Vento-tormo et al. 2018 Nature}.
It contains the 726 EVT, 953 dS2, 788 dNK1, and 533 dM1 cells.
}
\keyword{datasets}
