% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decidua}
\alias{decidua}
\title{Decidua Single Cell RNA-seq Dataset}
\format{
A sparse matrix with logTPM normalized transcript counts of 947 rows (ligand/receptor genes) and 3000 columns (cells).
}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6701/}
}
\usage{
decidua
}
\description{
A subset of log normalized decidua scRNA-seq dataset generated by \href{https://www.nature.com/articles/s41586-018-0698-6}{Vento-tormo et al. 2018 Nature}.
It contains the 726 EVT, 953 dS2, 788 dNK1, and 533 dM1 cells.
}
\concept{data}
\keyword{datasets}
