% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Differing}
\alias{Differing}
\title{Differential Interactions}
\usage{
Differing(
  object,
  interaction1,
  interaction2 = NULL,
  diff.thresh = 1,
  p.val.cutoff = 0.05,
  p.adjust.method = "none"
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{interaction1}{Cell-Cell interactions of interests.}

\item{interaction2}{Cell-Cell interactions to compare against. Default is NULL, all other cell types except the ones in interaction1 will be used.}

\item{diff.thresh}{Differential interaction strength cutoff, default is 1.}

\item{p.val.cutoff}{p value cutoff. Default is 0.05}

\item{p.adjust.method}{p value adjustment method. Default is "none". See details \code{\link[stats]{p.adjust}}.}
}
\value{
A data.frame with interaction differences and p values.
}
\description{
Conduct differential interaction analysis based on permutation similar to \code{\link[scMatchmaker]{Matchmaking}}.
To use this feature, save.perm parameter in \code{\link[scMatchmaker]{Matchmaking}} has to be set to TRUE when running Matchmaking.
}
\examples{
\dontrun{
object <- Differing(object, interaction1 = "Macrophage|CAF", interaction2 = "Macrophage|Tumor")
}
}
\concept{matchmake}
