% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filterInteractions}
\alias{filterInteractions}
\title{Filter Interaction Data}
\usage{
filterInteractions(
  interact.strength,
  interact.pval,
  interaction.meta,
  pval.cutoff = 0.05,
  strength.pct = 0.1,
  filter.cells = TRUE
)
}
\arguments{
\item{interact.strength}{Interaction strength matrix.}

\item{interact.pval}{Interaction p-value matrix.}

\item{interaction.meta}{Interaction metadata with first two columns as interaction partners matching interact.strength and interact.pval data.}

\item{pval.cutoff}{p value cutoff. Default is 0.05.}

\item{strength.pct}{Top strength quantile to select. Default is 0.1 aka the top 10\%.}

\item{filter.cells}{Whether to filter the cells. Default is TRUE.}
}
\value{
Returns a list containing filtered interaction strength, p value matrix and merged interaction data.
}
\description{
Filter interaction data based on interacting strength and p value cutoffs.
}
\keyword{internal}
