% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterMatrix}
\alias{filterMatrix}
\title{Filter Matrix}
\usage{
filterMatrix(matrix, low.cutoff = -Inf, high.cutoff = Inf)
}
\arguments{
\item{matrix}{A matrix to be filtered.}

\item{low.cutoff}{Lower cutoff bound below which will be removed. Default is negative infinity.}

\item{high.cutoff}{Upper cutoff bound above which will be removed. Default is positive infinity.}
}
\value{
Returns a filtered matrix.
}
\description{
Filter Matrix
}
\keyword{internal}
