% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{findInteractions}
\alias{findInteractions}
\title{Calculate the Interaction Strength}
\usage{
findInteractions(
  data,
  idents,
  ligand,
  receptor,
  stats = c("mean", "median"),
  pair.fxn = c("+", "*"),
  n_perm = 100,
  emd = FALSE,
  nbins = 100,
  weighted = TRUE,
  p.adjust.method = "none",
  threshold = 0,
  min_pct = 0,
  max_pct = 1,
  drop_zero = FALSE,
  cutoff.stats = NA,
  n_cores = NULL,
  seed = 1
)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{idents}{Cell type identity in characters or factors.}

\item{ligand}{Name or index for ligand.}

\item{receptor}{Name or index for receptor.}

\item{stats}{Method used to calculate strength.
\itemize{
  \item mean, using mean (Default).
  \item medain, using median.
  }}

\item{pair.fxn}{Functions used to calculate pair strength.
\itemize{
  \item +, using sum of the pairs (Default).
  \item *, using product of the pairs.
  }}

\item{n_perm}{Number of random permutations. Default is 100.}

\item{emd}{Whethter to run Earth Mover's Distance adjustment. Default is FALSE.}

\item{nbins}{Number of bins to use for Earth Mover's Distance calculation. Default is 10.}

\item{weighted}{Logic, compute weighted Earth Mover's Distance which will give higher weight/penalty on lowly expressed genes. Default is TRUE.}

\item{p.adjust.method}{p value adjustment method. Default is "none". See details \code{\link[stats]{p.adjust}}.}

\item{threshold}{Threshold used for interaction products. See details \code{\link[scMatchmaker]{calcStrength}}.}

\item{min_pct}{Earth Movre's Distance values below this percentile will be replaced with this percentile value. Default is 0.}

\item{max_pct}{Earth Movre's Distance Values above this percentile will be replaced with this percentile value. Default is 1.}

\item{drop_zero}{Whether to drop all the zeros in the data when calculating EMD. Default is FALSE.}

\item{cutoff.stats}{Cutoff used to calculate the statistics (mean or median). Default is NA. If set to 0, all 0 values will be removed when calculating mean or median.}

\item{n_cores}{Number of cores used. Default is to use all cores - 1. See details \code{\link[parallel]{makeCluster}}.}

\item{seed}{Random seed number for permutation. Default is 1.}
}
\value{
Returns a list contains interaction strength and pvalues, optionally the null strength and Earth Mover's Distance similarity.
}
\description{
Calculate the interaction strength and statistical significance using random permutation tests.
}
\keyword{internal}
