% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{findInteractions.single}
\alias{findInteractions.single}
\title{Calculate the Interaction Strength for Single Interaction}
\usage{
findInteractions.single(
  data,
  idents,
  ligand,
  receptor,
  stats = c("mean", "median"),
  pair.fxn = c("+", "*"),
  threshold = 0,
  cutoff = NA
)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{idents}{Cell type identity in characters or factors.}

\item{ligand}{Name or index for ligand.}

\item{receptor}{Name or index for receptor.}

\item{stats}{Method used to calculate strength.
\itemize{
  \item mean, using mean (Default).
  \item medain, using median.
  }}

\item{pair.fxn}{Functions used to calculate pair strength.
\itemize{
  \item +, using sum of the pairs (Default).
  \item *, using product of the pairs.
  }}

\item{cutoff}{Cutoff used to calculate the statistics (mean or median). Default is NA. If set to 0, all 0 values will be removed when calculating mean or median.}
}
\value{
Returns Earth Mover's Distance.
}
\description{
Calculate the Interaction Strength for a single interaction pairs.
}
\keyword{internal}
