% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{geomSketch}
\alias{geomSketch}
\title{Run Geometric Sketching}
\usage{
geomSketch(
  data,
  geom_size = 1000,
  n_pca = 10,
  scale = TRUE,
  seed = 1,
  which_python = Sys.which(names = "python3")
)
}
\arguments{
\item{data}{An M x d sparse matrix or data.frame with M rows of data points and d columns of features.}

\item{geom_size}{Size of geometric sketches to return. Default is 1000.}

\item{n_pca}{Number of PCA dimensions to use. Default is 10.}

\item{scale}{Whether to scale the data. Default is TRUE.}

\item{seed}{Random seed number for PCA. Default is 1.}

\item{which_python}{Path to python3 used.}
}
\value{
Returns geometric sketch IDs.
}
\description{
Run Geometric sketching sampling.
See \href{https://www.cell.com/cell-systems/pdfExtended/S2405-4712(19)30152-8}{Hie et al. 2019 Cell Systems} for details.
}
\keyword{internal}
