% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Generate Heatmap}
\usage{
heatmapPlot(
  data,
  which.cell = NULL,
  which.gene = NULL,
  aggregate = c("none", "row", "column"),
  counted = FALSE,
  color = "RdBu",
  color.bins = 7,
  order = TRUE,
  order.dist = "minkowski",
  p = 1,
  filter = FALSE,
  low.cutoff = -Inf,
  high.cutoff = Inf,
  return.data = FALSE,
  Rowv = NA,
  Colv = NA,
  trace = "none",
  margins = c(10, 10),
  na.color = "black",
  key = TRUE,
  show.text = TRUE,
  ...
)
}
\arguments{
\item{data}{Data matrix to plot}

\item{which.cell}{Interacting cell type name or index to plot. Default is NULL.}

\item{which.gene}{Interacting gene name or index to plot. Default is NULL.}

\item{aggregate}{Type of aggregation used.
\itemize{
  \item row, aggregated by row (Default).
  \item column, aggregated by column
  \item none, no aggregation.
  }}

\item{counted}{Aggregate interactions counts instead of strengths. Default is FALSE.}

\item{color}{Color palette name. Default is "RdBu", high values are in red and low values in blue. See details \code{\link[RColorBrewer]{brewer.pal}}.}

\item{color.bins}{Number of differen colors in pallete. Default is 7. See detail \code{\link[RColorBrewer]{brewer.pal}}.}

\item{order}{Whether to order the input data. Default is TRUE. See details \code{\link[seriation]{seriate}}.}

\item{order.dist}{Distance measure to be used. Default is "minkowski". See details \code{\link[stats]{dist}}.}

\item{p}{Power of Minkowski distance Default is 1, aka Manhattan distance. See details \code{\link[stats]{dist}}.}

\item{filter}{Whether to filter the input data based on low.cutoff and high.cutoff. Default is FALSE.}

\item{low.cutoff}{Lower cutoff bound below which will be removed. Default is negative infinity.}

\item{high.cutoff}{Upper cutoff bound above which will be removed. Default is positive infinity.}

\item{return.data}{Whether to return the data used to plot. Default is FALSE.}

\item{Rowv}{Whether to hierarchical cluster the rows. Default is NA. See details \code{\link[gplots]{heatmap.2}}.}

\item{Colv}{Whether to hierarchical cluster the columns Default is NA. See details \code{\link[gplots]{heatmap.2}}.}

\item{trace}{Whether to show the trace line. Defaults is none. See details \code{\link[gplots]{heatmap.2}}.}

\item{margins}{Plot margins. Defaults is c(10,10) See details \code{\link[gplots]{heatmap.2}}.}

\item{na.color}{Color for missing values (NAs). Defaults is black. See details \code{\link[gplots]{heatmap.2}}.}

\item{key}{Whether to show the key. Defaults is TRUE. See details \code{\link[gplots]{heatmap.2}}.}

\item{show.text}{Whether to show title, x-axis and y-axis text. Defaults is TRUE.}

\item{...}{Additioanl arguments passed to \code{\link[gplots]{heatmap.2}}.}
}
\value{
Plot a heatmap and optionally return the data used to generate the plot.
}
\description{
Generate a heatmap plot.
}
\keyword{internal}
