% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{histogramPlot}
\alias{histogramPlot}
\title{Generate Interaction Histograms}
\usage{
histogramPlot(
  data,
  idents,
  ligand,
  receptor,
  ligand.ident,
  receptor.ident,
  breaks = NULL,
  nbins = 100,
  freq = FALSE,
  cols.use = c("pink", "lightblue"),
  alpha = 0.7,
  ...
)
}
\arguments{
\item{data}{Data matrix to plot}

\item{idents}{Cell type identity.}

\item{ligand}{Ligand name.}

\item{receptor}{Receptor name.}

\item{ligand.ident}{Cell identity to plot for ligand.}

\item{receptor.ident}{Cell identity to plot for receptor.}

\item{breaks}{A vector giving the breakpoints between histogram cells. Default is NULL.}

\item{nbins}{Number of bins to use for histogram when breaks option is NULL. Default is 100.}

\item{freq}{Histogram of frequencies. Default is FALSE.}

\item{cols.use}{Two color code used for ligand and receptor. Default is pink and lightblue.}

\item{alpha}{Color transparency. Default is 0.7}

\item{...}{Additional argument passing to \code{\link[graphics]{hist}}.}
}
\value{
Plot a histogram.
}
\description{
Generate a histogram for a pair of ligand and receptor.
}
\keyword{internal}
