% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{LoadCellPhoneDB}
\alias{LoadCellPhoneDB}
\title{Load CellPhoneDB Datasets}
\usage{
LoadCellPhoneDB(
  interaction_input,
  gene_input,
  complex_input,
  gene.symbol = "gene_name",
  annotation.strategy = NULL,
  invert = FALSE,
  path = FALSE,
  url = FALSE
)
}
\arguments{
\item{interaction_input}{Data frame (csv file), directory path or URL of the interaction_input file.}

\item{gene_input}{Data frame (csv file), directory path or URL of the gene_input file.}

\item{complex_input}{Data frame (csv file), directory path or URL of the complex_input file.}

\item{gene.symbol}{Gene symbols to use. The possible names are listed as column names in the gene_input file.
Examples include: gene_name, uniprot, hgnc_symbol, ensembl. Default is gene_name.}

\item{annotation.strategy}{Annotation strategy used. Default is NULL, possible options include "curated" etc.}

\item{invert}{Whether to perfrom invert selection on annotation strategy. Default is FALSE}

\item{path}{Whether inputs are in directory path format.}

\item{url}{Whether inputs are in URl format.}
}
\value{
Returns the interaction pairs data.frame with the first two columns correspond to interacting partners.
}
\description{
Load the \href{https://www.cellphonedb.org/}{CellPhoneDB} dataset for ligand-receptor interaction pairs.
}
\examples{
\dontrun{
# CellPhoneDB URLs 
pair.data <- LoadCellPhoneDB(
 interaction_input = interaction_input.url,
 gene_input = gene_input.url,
 complex_input = complex_input.url,
 gene.symbol = "ensembl", annotation.strategy = "curated", url = TRUE)
}
}
\concept{processing}
