% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logCommand}
\alias{logCommand}
\title{Log Command}
\usage{
logCommand(pos = 1, frame = 1)
}
\arguments{
\item{pos}{Argument positions to skip. Default is 1.}

\item{frame}{Number of parental frame to go back. Default is 1. See \code{\link[base]{sys.frame}} for details.}
}
\value{
Returns a list of command arguments used.
}
\description{
Log Command
}
\keyword{internal}
