% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logTPM}
\alias{logTPM}
\title{Log-transformed Transcripts Per Million (logTPM)}
\usage{
logTPM(mat, normalize_factor = 10000)
}
\arguments{
\item{mat}{Count matrix. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{normalize_factor}{Normalization factor used. Default is 10000.}
}
\value{
Returns a regular or sparse matrix.
}
\description{
logTPM normalization using log base of 2.
}
\keyword{internal}
