% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmaker-class.R
\docType{class}
\name{Matchmaker-class}
\alias{Matchmaker-class}
\title{The Matchmaker S4 Class}
\description{
The Matchmaker object with the slot information listed as follow:
}
\section{Slots}{

\describe{
\item{\code{data}}{Expression data. A d x M matrix with d rows of interaction genes and M columns of data points (cells).}

\item{\code{annotation}}{Data frame of cell type annotatiuons.
\itemize{
  \item first column, stores the cell type identities for each cell in the data.
  \item ..., additional slots for future use.
  }}

\item{\code{interaction}}{Ligand-receptor interaction pairs. A dataframe with at least two columns for ligand-receptor pair gene names.}

\item{\code{strength}}{Matrix of relative strengths of interactions.
\itemize{
  \item row names, interacting cell types seperated by "|". 
  \item column names, interacting partner pairs seperated by "_".
  }}

\item{\code{pvalue}}{Matrix of p values of interaction strengths. Row and column naming are the same as the strength matrix.}

\item{\code{selected}}{A list of selected and filtered interactions:
\itemize{
  \item stregth, Matrix of relative strengths of interactions.
  \item pvalue, Matrix of p values of interaction strengths.
  \item merged.data, merged interaction pair data and strength data.
  }}

\item{\code{project_name}}{Name of the project.}

\item{\code{misc}}{List of miscellaneous:
\itemize{
  \item raw_data, raw expression data.
  \item stats_null, null interaction stength without earth mover's distance adjustment if emd=TRUE option is used.
  \item emd, earth mover's distance similarity.
  \item permute_result, list of permutation results if saved.
  \item sketch_id, geometric sketching ID.
  \item graph, graph visualization.
  \item single_interaction, single gene-gene interaction table in the format of single interactions if Complexing function is called.
  \item single_strength, single gene-gene interaction strengths if Complexing function is called.
  \item single_pvalue, single gene-gene interaction p-values of interactions strengths if Complexing function is called.
  \item unmerged_strength, unmerged interaction strenghts if Merging function is called.
  \item unmerged_pvalue, unmerged interaction p-values if Merging function is called.
  \item ..., additional slots for future use.
  }}

\item{\code{command}}{Nested list of commands and scMatchmaker version used.}
}}

