% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Matchmaking}
\alias{Matchmaking}
\title{Run Matchmaker}
\usage{
Matchmaking(
  object,
  ident1 = NULL,
  ident2 = NULL,
  stats = c("mean", "median"),
  pair.fxn = c("+", "*"),
  emd = FALSE,
  nbins = 100,
  n_perm = 100,
  weighted = FALSE,
  p.adjust.method = "none",
  min_emd_pct = 0,
  max_emd_pct = 1,
  drop_zero = FALSE,
  cutoff.stats = NA,
  n_cores = NULL,
  seed = 1
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{ident1}{Cell type identity group 1. Default is NULL and all cell types will be used.}

\item{ident2}{Cell type identity group 2. Default is NULL and all cell types will be used.}

\item{stats}{Statistics used to calculate strength.
\itemize{
  \item mean, using mean (Default).
  \item medain, using median.
  }}

\item{pair.fxn}{Functions used to calculate pair strength.
\itemize{
  \item +, using sum of the pairs (Default).
  \item *, using product of the pairs.
  }}

\item{emd}{Whethter to run Earth Mover's Distance adjustment. Default is FALSE.}

\item{nbins}{Number of bins to use for Earth Mover's Distance calculation. Default is 100.}

\item{n_perm}{Number of random permutations. Default is 100.}

\item{weighted}{Logic, compute weighted Earth Mover's Distance which will give higher weight/penalty on lowly expressed genes. Default is FALSE.}

\item{p.adjust.method}{p value adjustment method. Default is "none". See details \code{\link[stats]{p.adjust}}.}

\item{min_emd_pct}{Earth Mover's Distance below this percentile will be replaced with this percentile value. Default is 0.}

\item{max_emd_pct}{Earth Mover's Distance above this percentile will be replaced with this percentile value. Default is 1.}

\item{drop_zero}{Whether to drop all the zeros in the data when calculating EMD. Default is FALSE.}

\item{cutoff.stats}{Cutoff used to calculate the statistics (mean or median). Default is NA. If set to 0, all 0 values will be removed when calculating mean or median.}

\item{n_cores}{Number of cores used. Default is to use all cores - 1. See details \code{\link[parallel]{makeCluster}}.}

\item{seed}{Random seed number for permutation. Default is 1.}
}
\value{
Returns a Matchmaker object.
}
\description{
Run matchmaker to Identify cell-cell interactions.
}
\examples{
\dontrun{
object <- Matchmaking(object, stats = "mean", emd = TRUE, n_perm = 100)
}
}
