% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchNames}
\alias{matchNames}
\title{Match the Names of Inteactions Pairs in Expression Data.}
\usage{
matchNames(data, pairs, partner_a = 1, partner_b = 2)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{pairs}{Interaction pair data with at least two columns indicating gene-gene interaction partners.}

\item{partner_a}{Column name or index for interaction partner a. Defaulte is 1, the first column in interaction pair data.}

\item{partner_b}{Column name or index for interaction partner b. Defaulte is 2, the second column in interaction pairs data.}
}
\value{
Returns a list with filtered expression data and interaction pairs.
}
\description{
Match and filter the expression data and interacton pairs.
}
\keyword{internal}
