% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeLists}
\alias{mergeLists}
\title{Merge Two List}
\usage{
mergeLists(a, b)
}
\arguments{
\item{a, b}{Two lists.}
}
\value{
Returns the merged list.
}
\description{
Merge twp lists by the name of their elements.
}
\keyword{internal}
