% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Merging}
\alias{Merging}
\title{Merge Interactions}
\usage{
Merging(
  object,
  strength_merge = c("max", "average", "min"),
  pval_merge = c("min", "average", "max")
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{strength_merge}{Method to combine complex interaction strength from its subunits. Default is max (maximum).
\itemize{
  \item max, maximum strength of directed interactions.
  \item average, average strength of directed interactions.
  \item min, minimum strength of directed interaction.
  }}

\item{pval_merge}{Method to combine complex interaction p value from its subunits. Default is min (minimum).
\itemize{
  \item min, minimum p value of directed interactions.
  \item average, average p value of directed interactions.
  \item max, maximum p value of directed interactions.
  }}
}
\value{
Returns a Matchmaker object.
}
\description{
Merge directed (one-way) interactions into undirected (two-way) interactions.
For example, before merging, DC1|dM1 and dM1|DC1 represents two different cell-cell interactions. 
After merging, they will be combined as one undirected cell-cell interaction between DC1 and dM1.
}
\examples{
\dontrun{
object <- Merging(object, strength_merge = "average", pval_merge = "max")
}
}
\concept{matchmake}
