% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{Normalization}
\alias{Normalization}
\title{Data Normalization}
\usage{
Normalization(
  counts,
  normalization = c("logTPM", "cosine", "none"),
  normalize_factor = 10000,
  zero_percent = 0.7,
  verbose = FALSE
)
}
\arguments{
\item{counts}{Data to normalize. An N x M matrix with N rows of features and M columns of data points.}

\item{normalization}{Normalization method used. Default is cosine.
\itemize{
  \item logTPM, logged Transcript Per Million normalization: feature counts for each data point are divided by the total sum of them. Then the data is multiplied by the scale.factor before taking a log-transformed by log(1+x) (Default).
  \item cosine, cosine nomalization: feature counts for each data point are divided by their L2.
  \item none, additional normalization is not performed.
  }}

\item{normalize_factor}{Normalization factor used with lognorm method. Default is 10000.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}

\item{verbose}{Whether to display a process bar for cosine normalization. Default is FALSE.}
}
\value{
Returns the normalized data.
}
\description{
Normalize count or other data.
}
\examples{
\dontrun{
normalized.data <- Normalization(data)
}
}
\concept{processing}
