% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderMatrix}
\alias{orderMatrix}
\title{Order Matrix}
\usage{
orderMatrix(matrix, order.dist = "minkowski", p = 1, return.order = FALSE)
}
\arguments{
\item{matrix}{A matrix.}

\item{order.dist}{Distance measure to be used. Default is "minkowski". See details \code{\link[stats]{dist}}.}

\item{p}{Power of Minkowski distance Default is 1, aka Manhattan distance. See details \code{\link[stats]{dist}}.}

\item{return.order}{Whether to return the matrix order. Default is FALSE.}
}
\value{
Returns a matrix or list containing ordered matrix.
}
\description{
Order Matrix
}
\details{
See \code{\link[seriation]{seriate}}.
}
\keyword{internal}
