% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Interaction Histograms}
\usage{
PlotHistogram(
  object,
  ligand,
  receptor,
  ligand.ident,
  receptor.ident,
  fix.breaks = TRUE,
  nbins = 100,
  freq = FALSE,
  cols.use = c("pink", "lightblue"),
  alpha = 0.7,
  ...
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{ligand}{Ligand name.}

\item{receptor}{Receptor name.}

\item{ligand.ident}{Cell identity to plot for ligand.}

\item{receptor.ident}{Cell identity to plot for receptor.}

\item{fix.breaks}{Whether to fix the \code{\link[graphics]{hist}} breaks using all the data. Default is TRUE.}

\item{nbins}{Number of bins to use for histogram. Default is 100.}

\item{freq}{Histogram of frequencies. Default is FALSE.}

\item{cols.use}{Two color code used for ligand and receptor. Default is pink and lightblue.}

\item{alpha}{Color transparency. Default is 0.7.}

\item{...}{Additional argument passing to \code{\link[graphics]{hist}}.}
}
\value{
Plot a histogram plot.
}
\description{
Plot a histogram for a pair of ligand and receptor.
}
\examples{
\dontrun{
 PlotHistogram(object, ligand = "CSF1", receptor = "CSF1R", 
 ligand.ident = "Tumor", receptor = "Macrophages")
}
}
