% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Interaction Network Plot}
\usage{
PlotNetwork(
  object,
  selected = FALSE,
  which.cell = NULL,
  which.gene = NULL,
  aggregate = c("cell", "gene"),
  counted = FALSE,
  filter = FALSE,
  low.cutoff = -Inf,
  high.cutoff = Inf,
  edge.color = "black",
  node.color = NULL,
  node.size = 5,
  layout = NULL,
  directed = TRUE,
  bidirectional = TRUE,
  arrows = TRUE,
  louvain = FALSE,
  legend.posit = "topright",
  legend.breaks = 5,
  legend.title = "Interaction Strength",
  ...
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{selected}{Use selected data if calculated. Default is FALSE.}

\item{which.cell}{Interacting cell type name (saperated by '|' i.e 'DC|Mac') or index to plot. Default is NULL.}

\item{which.gene}{Interacting gene name (saperated by '_' i.e 'CSF1_CSF1R') or index to plot. Default is NULL.}

\item{aggregate}{Type of aggregation used. Default is none.
\itemize{
  \item cell, aggregated by cell-cell interactions.
  \item gene, aggregated by ligand-receptor pairs.
  }}

\item{counted}{Aggregate interactions counts instead of strengths. Default is FALSE.}

\item{filter}{Whether to filter the input data based on low.cutoff and high.cutoff. Default is FALSE.}

\item{low.cutoff}{Lower cutoff bound below which will be removed. Default is negative infinity.}

\item{high.cutoff}{Upper cutoff bound above which will be removed. Default is positive infinity.}

\item{edge.color}{Edge color. Default is black.}

\item{node.color}{Node color. Default is NULL.}

\item{node.size}{Node size. Default is 5}

\item{layout}{Layout used in the plot. Default is NULL.}

\item{directed}{Whether edges are directed. Default is TRUE.}

\item{bidirectional}{Whether edges are bidirectional. Default is TRUE.}

\item{arrows}{Whether to draw arrows. Default is TRUE.}

\item{louvain}{Whether to perform louvain clustering. Default is FALSE.}

\item{legend.posit}{Legend position. Default is topright.}

\item{legend.breaks}{Number of intervals to show. Default is 5.}

\item{legend.title}{Legend title. Default is 'Interaction Strength'.}

\item{...}{Additioanl arguments passed to \code{\link[qgraph]{qgraph}}.}
}
\value{
Return a network plot of interactions.
}
\description{
Generate a network plot of interactions.
}
\examples{
\dontrun{
 PlotNetwork(object, aggregate = "cell")
}
}
