% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotScatter}
\alias{PlotScatter}
\title{Interaction Scatter Plot}
\usage{
PlotScatter(
  object,
  ident1,
  ident2,
  ligands,
  receptors,
  use_raw = TRUE,
  add.lines = FALSE,
  add.text = TRUE,
  background.col = "grey",
  ligand.col = "red",
  receptor.col = "blue",
  ligand.pch = 16,
  receptor.pch = 16,
  point.cex = 1,
  label.offset = 1,
  ligand.pos = 2,
  receptor.pos = 4,
  legend.pos = "topright",
  ...
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{ident1}{Cell identity 1 to plot.}

\item{ident2}{Cell identity 2 to plot.}

\item{ligands}{Ligand name.}

\item{receptors}{Receptor name.}

\item{use_raw}{Logic, use raw data or ligand-receptor data. Default is TRUE.}

\item{add.lines}{Logic, draw lines between ligand-receptor pairs. Default is FALSE.}

\item{add.text}{Logic, add text for ligands and receptors. Default is TRUE.}

\item{background.col}{Background points color. Default is gray.}

\item{ligand.col}{Ligand points color. Default is red.}

\item{receptor.col}{Receptor points color. Default is blue.}

\item{ligand.pch}{Point shape for ligands. Default is 16 (solid circle).}

\item{receptor.pch}{Point shape for receptor. Default is 16 (solid circle).}

\item{point.cex}{Point size. Default is 1.}

\item{label.offset}{Text label offset. Default is 1.}

\item{ligand.pos}{Ligand label position. Default is 2 (left). See \code{\link[graphics]{text}}.}

\item{receptor.pos}{Receptor label position. Default is 4 (right). See \code{\link[graphics]{text}}.}

\item{legend.pos}{Legend position. Default is 'topright'.}

\item{...}{Additioanl arguments passed to \code{\link[base]{plot}}.}
}
\value{
Return a scatter plot of interactions.
}
\description{
Generate a scatter plot of interactions.
}
\examples{
\dontrun{
 PlotScatter(object, ident1 = "Tumor", ident2 = "Immune", ligands = "CD274", receptors = "PDCD1")
}
}
