% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randomIdents}
\alias{randomIdents}
\title{Generate Randomized Cell Identities}
\usage{
randomIdents(idents, n = 100, seed = 1)
}
\arguments{
\item{idents}{Cell type identity in characters or factors.}

\item{n}{Number of random trials. Default is 100.}

\item{seed}{Random seed number. Default is 1.}
}
\value{
Returns a matrix with randomized cell identities.
}
\description{
Generate Randomized Cell Identities
}
\keyword{internal}
