% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Resetting}
\alias{Resetting}
\title{Reset Interactions}
\usage{
Resetting(object, by = c("complex", "merge"))
}
\arguments{
\item{object}{Matchmaker object.}

\item{by}{Method to reset interaction data. Default is complex.
\itemize{
  \item complex, reset to single gene-gene interactions before Complexing is called.
  \item merge, reset to unmerged (directed) interactions before Merging is called.
  }}
}
\value{
Returns a Matchmaker object.
}
\description{
Reset complex-complex interactions back to single gene-gene interactions.
The single gene-gene interactions are stored in the @misc slot with names: single_interaction, single_strength, single_pvalue.
}
\examples{
\dontrun{
# reset to single gene-tene interaction
object <- Resetting(object, by = "complex")
# reset to unmerged interaction
object <- Resetting(object, by = "merge")
}
}
\concept{matchmake}
