% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowScale}
\alias{rowScale}
\title{Scale Data by Row}
\usage{
rowScale(mat, scale_max = 10)
}
\arguments{
\item{mat}{A matrix.}

\item{scale_max}{Maximum cutoff for scaled data. Default is 10.}
}
\value{
Returns a matrix.
}
\description{
Center and scale the expression matrix by its rows (features).
}
\keyword{internal}
