% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Saving}
\alias{Saving}
\title{Save Interaction Data}
\usage{
Saving(object, file_name, dir = NULL, selected = FALSE)
}
\arguments{
\item{object}{Matchmaker object.}

\item{file_name}{File name prefix to save as.}

\item{dir}{Directory. Default is current directory (NULL).}

\item{selected}{Use selected data if calculated. Default is FALSE.}
}
\value{
Return three csv files:
\itemize{
  \item filename_strength.csv, interaction strength data.
  \item filename_pval.csv, interaction p-value data.
  \item filename_rank.csv, ranked list of interactions.
  }
}
\description{
Save the interactions calculated.
}
\examples{
\dontrun{
Saving(object, file_name = "decidua")
}
}
