% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{Scaling}
\alias{Scaling}
\title{Data Scaling}
\usage{
Scaling(mat, scale_max = 10)
}
\arguments{
\item{mat}{Data matrix.}

\item{scale_max}{Maximum cutoff for scaled data. Default is 10.}
}
\value{
Returns a scaled and centered data.
}
\description{
Scale and center data.
}
\examples{
\dontrun{
scaled.data <- Scaling(data)
}
}
\keyword{internal}
