% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{scMatchmaker-package}
\alias{scMatchmaker-package}
\alias{_PACKAGE}
\title{\code{scMatchmaker} package}
\description{
scMatchmaker is an R package that uncovers cell-cell interactions in single cell RNA-seq data. 
It uses Earth Mover's Distance of ligand-receptor expressions and a permutation test to identify significant interactions in the data.  
It allows users to input customized interaction information or load the CellPhoneDB ligand-receptor database, 
and provides essential visualization tools.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/stevexniu/scMatchmaker/issues}
}

}
\author{
\strong{Maintainer}: Steve X. Niu \email{niuxsteve@gmail.com}

}
