% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{scMatchmaker}
\alias{scMatchmaker}
\title{\code{scMatchmaker} package}
\description{
scMatchmaker is an R package that uncovers cell-cell interactions in single cell RNA-seq data. 
It uses Earth Mover's Distance of ligand-receptor expressions and a permutation test to identify significant interactions in the data.  
It allows users to input customized interaction information or load the CellPhoneDB ligand-receptor database, 
and provides essential visualization tools.
}
\details{
See \href{https://github.com/stevexniu/scMatchmaker/blob/master/vignettes/vignette.Rmd}{vignette}  for an introduction to the package.
}
