% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Screening}
\alias{Screening}
\title{Screen for Interactions and Initialize Matchmaker Object}
\usage{
Screening(
  data,
  annotation,
  interaction,
  partner_a = 1,
  partner_b = 2,
  path = FALSE,
  project_name = "",
  zero_percent = 0.7
)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{annotation}{Cell type identity in following formats:
\itemize{
  \item character, cell type identity with either names as cell ids (barcodes) or in the same order matching the column names in data.
  \item data.frame, cell ids (barcodes) in row names and cell type identity in first column.
  }}

\item{interaction}{Interaction partner pair data with at least two columns indicating gene-gene interaction partners.}

\item{partner_a}{Column name or index for interaction partner a. Defaulte is 1, the first column in interaction pair data.}

\item{partner_b}{Column name or index for interaction partner b. Defaulte is 2, the second column in interaction pairs data.}

\item{path}{Whether inputs are in directory path format.}

\item{project_name}{Name of the project.}

\item{zero_percent}{Zero-entry percentage threshold. If the number of zero entries in the returned matrices is above this number, a sparse matrix will be returned. Default is 0.7 aka 70\%.}
}
\value{
Returns a Matchmaker object.
}
\description{
Screen for interactiins and initialize a Matchmaker object.
}
\examples{
\dontrun{
object <- Screening(data = data.use, annotation = idents.use, interaction = interaction.data)
}
}
