% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Selecting}
\alias{Selecting}
\title{Filter and Select Interaction Data}
\usage{
Selecting(object, pval.cutoff = 0.05, strength.pct = 0.1, filter.cells = TRUE)
}
\arguments{
\item{object}{Matchmaker object.}

\item{pval.cutoff}{p value cutoff. Default is 0.05.}

\item{strength.pct}{Top strength quantile to select. Default is 0.1 aka the top 10\%.}

\item{filter.cells}{Whether to filter the cells. Default is TRUE.}
}
\value{
Returns a Matchmaker object with filtered interactions saved in @selected slot.
}
\description{
Filter interaction using interacting strength and p value cutoffs. 
Non-significant interactions will be either removed or set to zero.
}
\examples{
\dontrun{
object <- Selecting(object, pval.cutoff = 0.05, strength.pct = 0.1)
}
}
