% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Sketching}
\alias{Sketching}
\title{Downsampling and Geometric Sketching}
\usage{
Sketching(
  object,
  cell_id = NULL,
  downsampling = TRUE,
  size = 2000,
  geom_pca_dims = 10,
  scale = TRUE,
  seed = 1,
  which_python = Sys.which(names = "python3")
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{cell_id}{Cell IDs used to subset the data. Default is NULL. If provided the downsampling will be skipped.}

\item{downsampling}{Perform downsampling based on the idents probability rather than geometric sketching. Default is TRUE.}

\item{size}{Number of cells to return. Default is 2000. Please note that if downsampling is performed the actual size may vary due to rounding effect.}

\item{geom_pca_dims}{Number of PCA dimensions to use. Default is 10.}

\item{scale}{Whether to scale the data. Default is TRUE.}

\item{seed}{Random seed number for sampling. Default is 1.}

\item{which_python}{Path to python3 used.}
}
\value{
Returns a Matchmaker object.
}
\description{
Run downsampling or geometric sketching to reduce number of cells for interaction calculation.
See \href{https://www.cell.com/cell-systems/pdfExtended/S2405-4712(19)30152-8}{Hie et al. 2019 Cell Systems} for details.
}
\examples{
\dontrun{
object <- Sketching(object, size = 2000)
}
}
\concept{matchmake}
