% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{statsCells}
\alias{statsCells}
\title{Calculate the Statistics for each Cell.}
\usage{
statsCells(data, idents, stats = c("mean", "median"), cutoff = NA)
}
\arguments{
\item{data}{Expression data. A d x M matrix with d rows of features and M columns of data points (cells).}

\item{idents}{Cell type identity in characters or factors.}

\item{stats}{Statistics used to calculate strength.
\itemize{
  \item mean, using mean (Default).
  \item medain, using median.
  }}

\item{cutoff}{Cutoff used to calculate the statistics (mean or median). Default is NA. If set to 0, all 0 values will be removed when calculating mean or median.}
}
\value{
Returns a matrix with calculated statistics.
}
\description{
Calculate the statistics, mean or median, for each cell.
}
\keyword{internal}
