% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchmake.R
\name{Subsetting}
\alias{Subsetting}
\title{Subset Interactions}
\usage{
Subsetting(
  object,
  ident1 = NULL,
  ident2 = NULL,
  partner_a = NULL,
  partner_b = NULL,
  directed = FALSE
)
}
\arguments{
\item{object}{Matchmaker object.}

\item{ident1}{Cell type identity group 1 to subset.}

\item{ident2}{Cell type identity group 2 to subset. Default is NULL, and all cell types will be used.}

\item{partner_a}{Interaction partner a (See \code{\link[scMatchmaker]{Screening}}) to subset. Defaulte is NULL, and all interactions will be selected.}

\item{partner_b}{Interaction partner b (See \code{\link[scMatchmaker]{Screening}}) to subset. Defaulte is NULL, and all interactions will be selected.}

\item{directed}{Logic, if TRUE subset only ident1 with ligands (partner_a) and ident2 with receptor (partner_b). 
If FALSE, ident2 with ligands and ident1 with receptor will be included. Default is TRUE.}
}
\value{
Returns a new Matchmaker object.
}
\description{
Subset the interactions based on cell type identity and interacting partners.
}
\examples{
\dontrun{
# subset all Macrophage related interactions
interaction.mac <- Subsetting(object, ident1 = "Macrophage")
# subset all CSF1 related interactions
interaction.csf <- Subsetting(object, partner_a = "CSF1")
}
}
\concept{matchmake}
