% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroOne}
\alias{zeroOne}
\title{Convert to Zero and One Range}
\usage{
zeroOne(x, min_pct = 0, max_pct = 1)
}
\arguments{
\item{x}{Data.}

\item{min_pct}{Values below this percentile will be replaced with this percentile value. Default is 0.}

\item{max_pct}{Values above this percentile will be replaced with this percentile value. Default is 1.}
}
\value{
Returns new values in zero and one range.
}
\description{
Convert data to values between zero and one.
}
\keyword{internal}
